requirements.txt
tornado>=2.2.1
pycassa>=1.6.0
Mako>=0.7.0
celery-with-redis>=2.4.1